requirements.txt
flask 
openai 
werkzeug
python-dotenv